*: Refactor package organization
1. Move evaluator from optimizer to root dir.
2. Move builtin from expression to evaluator.
3. Remove expression package.
