+++ evaluator/builtin/string_test.go
